util/time: move more functions
