+++ src/util/time.rs
@@ -15,6 +15,31 @@ use std::time::{SystemTime, Duration};
 use std::thread::{self, JoinHandle, Builder};
 use std::sync::mpsc::{self, Sender};
 
+/// Convert Duration to milliseconds.
+#[inline]
+pub fn duration_to_ms(d: Duration) -> u64 {
+    let nanos = d.subsec_nanos() as u64;
+    // Most of case, we can't have so large Duration, so here just panic if overflow now.
+    d.as_secs() * 1_000 + (nanos / 1_000_000)
+}
+
+/// Convert Duration to seconds.
+#[inline]
+pub fn duration_to_sec(d: Duration) -> f64 {
+    let nanos = d.subsec_nanos() as f64;
+    // Most of case, we can't have so large Duration, so here just panic if overflow now.
+    d.as_secs() as f64 + (nanos / 1_000_000_000.0)
+}
+
+/// Convert Duration to nanoseconds.
+#[inline]
+pub fn duration_to_nanos(d: Duration) -> u64 {
+    let nanos = d.subsec_nanos() as u64;
+    // Most of case, we can't have so large Duration, so here just panic if overflow now.
+    d.as_secs() * 1_000_000_000 + nanos
+}
+
+
 const DEFAULT_WAIT_MS: u64 = 100;
 
 pub struct Monitor {
@@ -84,6 +109,7 @@ mod tests {
     use std::time::{SystemTime, Duration};
     use std::thread;
     use std::ops::Sub;
+    use std::f64;
     use super::*;
 
     use std::sync::atomic::{AtomicBool, Ordering};
@@ -111,5 +137,18 @@ mod tests {
         thread::sleep(Duration::from_secs(1));
 
         assert_eq!(jumped.load(Ordering::SeqCst), true);
+    }
+
+    #[test]
+    fn test_duration_to() {
+        let tbl = vec![0, 100, 1_000, 5_000, 9999, 1_000_000, 1_000_000_000];
+        for ms in tbl {
+            let d = Duration::from_millis(ms);
+            assert_eq!(ms, duration_to_ms(d));
+            let exp_sec = ms as f64 / 1000.0;
+            let act_sec = duration_to_sec(d);
+            assert!((act_sec - exp_sec).abs() < f64::EPSILON);
+            assert_eq!(ms * 1_000_000, duration_to_nanos(d));
+        }
     }
 }
